requirements.txt
mysql-connector-python
bcrypt
requests